+++ src/ast/formula.rs
@@ -1,4 +1,5 @@
 use crate::*;
+use std::fmt::Display;
 
 #[derive(Debug, Clone)]
 pub enum Goal {
@@ -26,4 +27,36 @@ pub enum Prog {
     And(Vec<Prog>),
     ForAll(Vec<IdentSort>, Box<Prog>),
     Implies(PrimGoal, Box<Prog>),
+    // Exists via skolemization
+    // Exists via
+}
+
+impl Display for Prog {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Prog::Atom(a) => std::fmt::Display::fmt(a, f),
+            Prog::And(ps) => write!(f, "(and {})", ListDisplay(ps, " ")),
+            Prog::ForAll(xs, p) => write!(f, "(forall ({}) {})", ListDisplay(xs, " "), *p),
+            Prog::Implies(gs, p) => write!(f, "(=> {} {})", gs, *p),
+        }
+    }
+}
+
+impl Display for PrimGoal {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            PrimGoal::Atom(fact) => write!(f, "{}", fact),
+            PrimGoal::And(ps) => write!(f, "(and {})", ListDisplay(ps, " ")),
+        }
+    }
+}
+
+impl Display for Goal {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match self {
+            Goal::PrimGoal(pg) => write!(f, "{}", pg),
+            Goal::ForAll(xs, p) => write!(f, "(forall ({}) {})", ListDisplay(xs, " "), *p),
+            Goal::Implies(gs, p) => write!(f, "(=> {} {})", gs, *p),
+        }
+    }
 }
