+++ src/lib.rs
@@ -1,12 +1,12 @@
 pub mod ast;
 mod extract;
 mod gj;
+mod logic;
 pub mod sort;
 mod typecheck;
 mod unionfind;
 pub mod util;
 mod value;
-
 use hashbrown::hash_map::Entry;
 use indexmap::map::Entry as IEntry;
 use instant::{Duration, Instant};
@@ -31,7 +31,6 @@ type ArcSort = Arc<dyn Sort>;
 pub use value::*;
 
 use gj::*;
-
 use unionfind::*;
 use util::*;
 
@@ -943,10 +942,10 @@ impl EGraph {
             }
             Command::Prove(goal) => {
                 self.backchain_goal(goal.clone())?;
-                format!("Proof succeeded: {:?}", goal)
+                format!("Proof succeeded: {}", goal)
             }
             Command::Assert(prog) => {
-                self.assert_prog(&mut vec![], prog.clone())?;
+                self.assert_prog(prog.clone())?;
                 format!("Asserted {:?}", prog)
             }
             Command::Extract { e, variants } => {
@@ -967,9 +966,9 @@ impl EGraph {
             Command::Check(fact) => {
                 if should_run {
                     self.check_fact(&fact, true)?;
-                    "Checked.".into()
+                    format!("Checked {fact}.")
                 } else {
-                    "Skipping check.".into()
+                    format!("Skipping check {fact}.")
                 }
             }
             Command::Action(action) => {
@@ -991,7 +990,7 @@ impl EGraph {
             Command::Declare(IdentSort { name, sort }) => {
                 if should_run {
                     let arcsort = self.sorts.get(&sort).unwrap().clone();
-                    self.declare_const(name, &arcsort);
+                    self.declare_const(name, &arcsort)?;
                     format!("Declared {name}: {sort}")
                 } else {
                     format!("Skipping decare {name}")
@@ -1124,156 +1123,6 @@ impl EGraph {
         }
     }
 
-    fn calc_helper(
-        &mut self,
-        idents: Vec<IdentSort>,
-        exprs: Vec<Expr>,
-        depth: &mut i64,
-    ) -> Result<(), Error> {
-        self.push();
-        *depth += 1;
-        // Insert fresh symbols for locally universally quantified reasoning.
-        for IdentSort { name, sort } in idents {
-            let sort = self.sorts.get(&sort).unwrap().clone();
-            self.declare_const(name, &sort)?;
-        }
-        // Insert each expression pair and run until they match.
-        for ab in exprs.windows(2) {
-            let a = &ab[0];
-            let b = &ab[1];
-            self.push();
-            *depth += 1;
-            self.eval_expr(a, None, true)?;
-            self.eval_expr(b, None, true)?;
-            let cond = Fact::Eq(vec![a.clone(), b.clone()]);
-            self.run_command(
-                Command::Run(RunConfig {
-                    limit: 100000,
-                    until: Some(cond.clone()),
-                }),
-                true,
-            )?;
-            self.run_command(Command::Check(cond), true)?;
-            self.pop().unwrap();
-            *depth -= 1;
-        }
-        self.pop().unwrap();
-        *depth -= 1;
-        Ok(())
-    }
-
-    // Prove a sequence of equalities universally quantified over idents
-    pub fn calc(&mut self, idents: Vec<IdentSort>, exprs: Vec<Expr>) -> Result<(), Error> {
-        if exprs.len() < 2 {
-            Ok(())
-        } else {
-            let mut depth = 0;
-            let res = self.calc_helper(idents, exprs, &mut depth);
-            if res.is_err() {
-                // pop egraph back to original state if error
-                for _ in 0..depth {
-                    self.pop()?;
-                }
-            } else {
-                assert!(depth == 0);
-            }
-            res
-        }
-    }
-
-    fn backchain_primgoal(&mut self, goal: PrimGoal) -> Result<(), Error> {
-        match goal {
-            PrimGoal::Atom(fact) => {
-                if self.check_fact(&fact, true).is_err() {
-                    dbg!("here runnign");
-                    self.run_command(
-                        Command::Run(RunConfig {
-                            limit: 100000,
-                            until: Some(fact),
-                        }),
-                        true,
-                    )?;
-                }
-                Ok(())
-            }
-            PrimGoal::And(goals) => {
-                for goal in goals {
-                    self.backchain_primgoal(goal)?;
-                }
-                Ok(())
-            }
-        }
-    }
-    fn assert_prog(&mut self, body: &mut Vec<Fact>, prog: Prog) -> Result<(), Error> {
-        match prog {
-            Prog::Atom(action) => {
-                if body.len() == 0 {
-                    self.eval_actions(&vec![action])
-                } else {
-                    let _ = self.add_rule(ast::Rule {
-                        body: body.clone(),
-                        head: vec![action],
-                    })?; // Hmm duplicate rules error?
-                    Ok(())
-                }
-            }
-            Prog::And(progs) => {
-                for prog in progs {
-                    self.assert_prog(&mut body.clone(), prog)?;
-                }
-                Ok(())
-            }
-            Prog::ForAll(idents, prog) => {
-                for ident in idents {
-                    // This is fishy
-                    // I should actually track what is in the context.
-                    assert!(!self.functions.contains_key(&ident.name));
-                }
-                self.assert_prog(body, *prog)
-            }
-            Prog::Implies(goal, prog) => {
-                EGraph::body_from_primgoal(body, goal);
-                self.assert_prog(body, *prog)
-            }
-        }
-    }
-    fn body_from_primgoal(body: &mut Vec<Fact>, g: PrimGoal) {
-        match g {
-            PrimGoal::Atom(f) => body.push(f),
-            PrimGoal::And(goals) => {
-                for goal in goals {
-                    EGraph::body_from_primgoal(body, goal);
-                }
-            }
-        }
-    }
-    fn backchain_goal_helper(&mut self, goal: Goal) -> Result<(), Error> {
-        match goal {
-            Goal::ForAll(idents, goal) => {
-                for IdentSort { name, sort } in idents {
-                    let sort = self.sorts.get(&sort).unwrap().clone();
-                    self.declare_const(name, &sort)?;
-                }
-                self.backchain_goal_helper(*goal)
-            }
-            Goal::PrimGoal(goal) => {
-                dbg!("primgoal");
-                self.backchain_primgoal(goal)
-            }
-            Goal::Implies(prog, goal) => {
-                self.assert_prog(&mut vec![], prog)?;
-                self.backchain_goal_helper(*goal)
-            }
-        }
-    }
-    pub fn backchain_goal(&mut self, goal: Goal) -> Result<(), Error> {
-        self.push();
-        dbg!("backchainign");
-        let res = self.backchain_goal_helper(goal);
-        self.pop().unwrap();
-        res
-    }
-
     // Extract an expression from the current state, returning the cost, the extracted expression and some number
     // of other variants, if variants is not zero.
     pub fn extract_expr(
@@ -1364,6 +1213,17 @@ impl EGraph {
 
     pub fn num_tuples(&self) -> usize {
         self.functions.values().map(|f| f.nodes.len()).sum()
+    }
+
+    pub fn summary(&self) -> String {
+        let functions = ListDisplay(self.functions.keys().collect::<Vec<_>>(), ", ");
+        let rules = ListDisplay(self.rules.keys().collect::<Vec<_>>(), ", ");
+        format!(
+            "Egraph Summary:\n\
+            Functions: {functions}\n\
+            Rules: {rules}\n\
+            "
+        )
     }
 }
