+++ tests/proof.egg
@@ -1,0 +1,48 @@
+
+
+(datatype Vert);(a) (b) (c))
+
+(declare a Vert)
+(declare b Vert)
+(declare c Vert)
+(relation edge (Vert Vert))
+(relation path (Vert Vert))
+(print edge)
+;(check (edge (a) (c)))
+(print edge)
+(assert (forall ((x Vert) (y Vert)) (=> (edge x y) (path x y))))
+(assert (forall ((x Vert) (y Vert) (z Vert)) (=> (and (edge x y) (path y z)) (path x z))))
+(assert (edge a b))
+(assert (edge b c))
+;(run 10)
+(check (edge a b))
+(print path)
+(print edge)
+;(check (edge a c))
+;(check (path c a))
+(print path)
+(prove (path a b))
+(prove (path a c))
+; (prove (path c a))
+(print path)
+(clear)
+(clear-rules)
+; what is cleared?
+; ok they are all created. but unfilled
+(print path)
+(print edge)
+(print a)
+; (print d) fails
+(prove
+    (=>
+        (and  
+            (edge a b)
+            (edge b c)
+            (forall ((x Vert) (y Vert)) (=> (edge x y) (path x y)))
+            (forall ((x Vert) (y Vert) (z Vert)) (=> (and (edge x y) (path y z)) (path x z)))
+        )
+        (path a c)
+    )
+)
+
+
